+++ src/canvas/animationcanvas.h
@@ -13,8 +13,8 @@ namespace PixelMaestro {
 			AnimationCanvas(Section* section);
 			AnimationCanvas(Section* section, uint16_t num_frames);
 			~AnimationCanvas();
-			void activate(uint32_t pixel);
-			void deactivate(uint32_t pixel);
+			void activate(uint16_t x, uint16_t y);
+			void deactivate(uint16_t x, uint16_t y);
 			void draw_frame(bool* frame, uint16_t size_x, uint16_t size_y);
 			bool* get_frame(uint16_t frame);
 			Colors::RGB get_pixel_color(uint16_t x, uint16_t y);