+++ src/core/section.h
@@ -11,6 +11,7 @@
 #include "../canvas/canvas.h"
 #include "../canvas/canvastype.h"
 #include "colors.h"
+#include "maestro.h"
 #include "point.h"
 #include "pixel.h"
 
@@ -18,6 +19,8 @@ namespace PixelMaestro {
 	class Canvas;
 
 	class Animation;
+
+	class Maestro;
 
 	class Section {
 
@@ -63,22 +66,22 @@ namespace PixelMaestro {
 			Animation* get_animation();
 			Canvas* get_canvas();
 			Point* get_dimensions();
+			Maestro* get_maestro();
 			Section::Overlay* get_overlay();
 			uint8_t get_num_overlays();
 			Section* get_parent_section();
 			Pixel* get_pixel(uint16_t x, uint16_t y);
 			Colors::RGB get_pixel_color(uint16_t x, uint16_t y);
-			uint16_t get_refresh_interval();
 			void remove_animation();
 			void remove_canvas();
 			void remove_overlay();
 			Animation* set_animation(AnimationType::Type type, Colors::RGB* colors, uint8_t num_colors, bool preserve_cycle_index = true);
 			Canvas* set_canvas(CanvasType::Type type = CanvasType::Type::AnimationCanvas, uint16_t num_frames = 1);
 			void set_dimensions(uint16_t x, uint16_t y);
+			void set_maestro(Maestro* maestro);
 			void set_one(uint32_t pixel, Colors::RGB* color);
 			void set_one(uint16_t x, uint16_t y, Colors::RGB* color);
 			Section::Overlay* set_overlay(Colors::MixMode mix_mode = Colors::MixMode::Alpha, uint8_t alpha = 128);
-			void set_refresh_interval(uint16_t interval);
 			void update(const uint32_t& current_time);
 
 		private:
@@ -91,6 +94,9 @@ namespace PixelMaestro {
 			/// The logical layout of the Pixels.
 			Point dimensions_;
 
+			/// The Section's parent Maestro.
+			Maestro* maestro_ = nullptr;
+
 			/// The Section overlaying the current section (if applicable).
 			Overlay* overlay_ = nullptr;
 
@@ -99,9 +105,6 @@ namespace PixelMaestro {
 
 			/// The array of Pixels managed by the Section.
 			Pixel* pixels_ = nullptr;
-
-			/// The time between Pixel redraws in milliseconds. Dictated by the Maestro.
-			uint16_t refresh_interval_ = 50;
 	};
 }