+++ src/core/timing.cpp
@@ -7,36 +7,44 @@
 namespace PixelMaestro {
 	/**
 	 * Constructor. Sets the speed interval.
-	 * @param speed Amount of time (in milliseconds) between events.
+	 * @param interval Amount of time (in milliseconds) between events.
 	 */
-	Timing::Timing(uint16_t speed) {
-		set_speed(speed);
+	Timing::Timing(uint16_t interval) {
+		set_interval(interval);
 	}
 
 	/**
-	 * Returns the speed.
-	 * @return Speed.
+	 * Returns the interval between events.
+	 * @return Interval.
 	 */
-	uint16_t Timing::get_speed() {
-		return speed_;
+	uint16_t Timing::get_interval() {
+		return interval_;
+	}
+
+	/**
+	 * Returns the last update time.
+	 * @return Last update time.
+	 */
+	uint32_t Timing::get_last_time() {
+		return last_time_;
 	}
 
 	/**
 	 * Sets the amount of time between events.
 	 *
-	 * @param speed Amount of time (in milliseconds) between events.
+	 * @param interval Amount of time (in milliseconds) between events.
 	 */
-	void Timing::set_speed(uint16_t speed) {
-		this->speed_ = speed;
+	void Timing::set_interval(uint16_t interval) {
+		this->interval_ = interval;
 	}
 
 	/**
 	 * Checks timings.
 	 * @param current_time Current program runtime.
-	 * @return If the runtime exceeds speed_, return true.
+	 * @return If the runtime exceeds the interval, return true.
 	 */
 	bool Timing::update(const uint32_t& current_time) {
-		if ((current_time - last_time_) >= speed_) {
+		if ((current_time - last_time_) >= interval_) {
 			last_time_ = current_time;
 			return true;
 		}
@@ -47,11 +55,11 @@ namespace PixelMaestro {
 	// AnimationTiming Methods
 	/**
 	 * Constructor. Sets the speed and pause intervals.
-	 * @param speed Amount of time (in milliseconds) between animation cycles.
+	 * @param interval Amount of time (in milliseconds) between animation cycles.
 	 * @param pause Delay (in milliseconds) before starting the next animation cycle.
 	 */
-	AnimationTiming::AnimationTiming(uint16_t speed, uint16_t pause) {
-		set_speed(speed, pause);
+	AnimationTiming::AnimationTiming(uint16_t interval, uint16_t pause) {
+		set_interval(interval, pause);
 	}
 
 	/**
@@ -81,7 +89,7 @@ namespace PixelMaestro {
 		 * Otherwise, just jump to the next cycle.
 		 */
 		if (fade) {
-			step_count_ = (speed_ - pause_) / (float)refresh_interval;
+			step_count_ = (interval_ - pause_) / (float)refresh_interval;
 
 			// Make sure step_count_ is at least 1
 			if (step_count_ == 0) {
@@ -93,8 +101,14 @@ namespace PixelMaestro {
 		}
 	}
 
-	void AnimationTiming::set_speed(uint16_t speed, uint16_t pause) {
-		this->speed_ = speed;
+	/**
+	 * Sets the amount of time between events.
+	 *
+	 * @param interval Amount of time (in milliseconds) between events.
+	 * @param pause The amount of time (in milliseconds) to wait before starting an animation cycle.
+	 */
+	void AnimationTiming::set_interval(uint16_t interval, uint16_t pause) {
+		this->interval_ = interval;
 		this->pause_ = pause;
 	}
 }