+++ src/core/timing.h
@@ -11,8 +11,9 @@ namespace PixelMaestro {
 	class Timing {
 		public:
 			Timing(uint16_t speed = 1000);
-			uint16_t get_speed();
-			void set_speed(uint16_t speed);
+			uint16_t get_interval();
+			uint32_t get_last_time();
+			void set_interval(uint16_t interval);
 			bool update(const uint32_t& current_time);
 
 		protected:
@@ -20,16 +21,16 @@ namespace PixelMaestro {
 			uint32_t last_time_ = 0;
 
 			/// The amount of time (in milliseconds) between events.
-			uint16_t speed_;
+			uint16_t interval_;
 	};
 
 	class AnimationTiming : public Timing {
 		public:
-			AnimationTiming(uint16_t speed = 1000, uint16_t pause = 0);
+			AnimationTiming(uint16_t interval = 1000, uint16_t pause = 0);
 			uint16_t get_pause();
 			uint8_t get_step_count();
 			void recalculate_step_count(bool fade, uint16_t refresh_interval);
-			void set_speed(uint16_t speed, uint16_t pause = 0);
+			void set_interval(uint16_t interval, uint16_t pause = 0);
 
 		private:
 			/// The amount of time (in milliseconds) to wait before starting an animation cycle. Defaults to 0.